requirements.txt
Django==6.0.2
gunicorn==23.0.0
whitenoise==6.9.0
kafka-python==2.1.3
Pillow==11.2.1
cryptography>=42.0.0
django-ninja>=1.0.0
